Moved eth utils out into it's own package
